more duration logic to server
