+++ apps/client/src/common/hooks/useEntryAction.ts
@@ -49,10 +49,10 @@ import {
   requestEventSwap,
   requestGroupEntries,
   requestUngroup,
+  requestFitGroupTarget,
 } from '../api/rundown';
 import { logAxiosError } from '../api/utils';
 import { useEditorSettings } from '../stores/editorSettings';
-import { eventDurationMatchGroupTarget } from '../utils/time';
 
 export type EventOptions = Partial<{
   // options of any new entries (event / delay / group)
@@ -473,26 +473,16 @@ function useEntryActionsForRundown(scopedRundownId: string | undefined) {
   /**
    * Updates time of existing event so it satisfies the group target duration
    * @param eventId {EntryId} - id of the event
-   * @param groupId {EntryId} - id of the enclosing group
    */
   const matchGroupDuration = useCallback(
-    async (eventId: EntryId, groupId: EntryId) => {
-      const rundown = queryClient.getQueryData<Rundown>(resolveCurrentRundownQueryKey());
-      if (!rundown) return;
-      const group = rundown.entries[groupId];
-      if (!group || !isOntimeGroup(group)) return;
-      const event = rundown.entries[eventId];
-      if (!event || !isOntimeEvent(event)) return;
-
-      const newDuration = eventDurationMatchGroupTarget({
-        targetDuration: group.targetDuration,
-        groupDuration: group.duration,
-        eventDuration: event.duration,
-      });
-      if (!newDuration) return;
-      updateTimer(eventId, 'duration', String(newDuration / MILLIS_PER_SECOND) + 's', false);
+    async (eventId: EntryId) => {
+      const rundownId = getCurrentRundownData()?.id;
+      if (!rundownId) {
+        throw new Error('Rundown not initialised');
+      }
+      await requestFitGroupTarget(rundownId, eventId);
     },
-    [queryClient, updateTimer, resolveCurrentRundownQueryKey],
+    [getCurrentRundownData],
   );
 
   /**
